+++ isolde/src/restraints_cpp/position_restraints_ext.h
@@ -320,7 +320,7 @@ position_restraint_target_vector(void *restraint, size_t n, double *vec)
     }
 }
 
-extern "C" EXPORT void position_restraint_bond_transform(void *restraint, size_t n, double *transform)
+extern "C" EXPORT void position_restraint_bond_transform(void *restraint, size_t n, float *transform)
 {
     Position_Restraint **r = static_cast<Position_Restraint **>(restraint);
     try {
